+++ test/tlds.js
@@ -33,14 +33,15 @@ describe('#whoiser.tld()', function() {
 			assert.equal(whois.whois, 'whois.nic.google', 'WHOIS server doesn\'t match')
 		});
 
-		it('should return WHOIS for ".香港" - IDN', async function() {
-			let whois = await whoiser('.香港')
-			assert.equal(whois.domain, '香港', 'TLD doesn\'t match')
-			assert.equal(whois.whois, 'whois.hkirc.hk', 'WHOIS server doesn\'t match')
+		it('should return WHOIS for "analytics" (no whois server)', async function() {
+			let whois = await whoiser.tld('analytics')
+			assert.equal(whois.whois, 'whois.nic.analytics', 'WHOIS server doesn\'t match')
+			assert.equal(whois.domain, 'ANALYTICS', 'TLD doesn\'t match')
+			assert.equal(whois.created, '2015-11-20', 'Created date doesn\'t match')
 		});
 
-		it('should return WHOIS for ".XN--J6W193G" - IDN', async function() {
-			let whois = await whoiser('.XN--J6W193G')
+		it('should return WHOIS for ".香港" - IDN', async function() {
+			let whois = await whoiser('.香港')
 			assert.equal(whois.domain, '香港', 'TLD doesn\'t match')
 			assert.equal(whois.whois, 'whois.hkirc.hk', 'WHOIS server doesn\'t match')
 		});