+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Whoiser change log
 
+#### 1.15.0 - 4 February 2023
+
+- Updated - Improved .it parsing, preserve date structure and more [#92](https://github.com/LayeredStudio/whoiser/pull/92)
+- Updated - Merge/normalize more WHOIS domain labels
+- Fixed - Uses raw domain for .de TLD in `whoisDomain` [#95](https://github.com/LayeredStudio/whoiser/pull/95)
+- Fixed - Query more sources for TLD whois server [#86](https://github.com/LayeredStudio/whoiser/pull/86)
+
 #### 1.14.0 - 14 January 2023
 
 - Added - Support for third level domains [#80](https://github.com/LayeredStudio/whoiser/pull/80)